+++ .gitignore
@@ -146,4 +146,6 @@ lib
 .db/*
 !.db/.gitkeep
 
+.firebase/
+
 # End of https://www.toptal.com/developers/gitignore/api/node,visualstudiocode,yarn
